#69 rename: 임의의 폴더명으로 변경
