+++ components/editWiki/input/EditImageInput.tsx
