+++ components/listWiki/SPinner.tsx
