+++ components/myWiki/ProfileDetail.module.scss
